split api package into files, client request, compilation of server/client without cgo
